requirements.txt
psycopg2-binary
geoip2
